move incremental algorithm to closeness.py
