+++ networkx/algorithms/centrality/__init__.py
@@ -1,7 +1,6 @@
 from .betweenness import *
 from .betweenness_subset import *
 from .closeness import *
-from .incremental_closeness import *
 from .subgraph_alg import *
 from .current_flow_closeness import *
 from .current_flow_betweenness import *
